pyproject.toml
[tool.poetry]
name = "severity_estimation"
version = "0.1.0"
description = "Perception Failure Severity Estimation"
authors = ["Pasquale Antonante"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
poethepoet = "^0.14.0" # task manager
# nuplan-devkit = { git = "https://github.com/motional/nuplan-devkit.git", branch = "nuplan-devkit-v0.3" }
nuplan = {path = "../nuplan", develop = true}
hydra-core = "^1.2.0"
loguru = "^0.6.0"
copulae = "^0.7.7"
imageio = "^2.19.5"
hj-reachability = "^0.4.0"

# PyTorch
torch = { url = "https://download.pytorch.org/whl/cu111/torch-1.8.1%2Bcu111-cp39-cp39-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu111/torchvision-0.9.1%2Bcu111-cp39-cp39-linux_x86_64.whl" }

# JAX
# jaxlib = { url = "https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.3.25+cuda11.cudnn82-cp39-cp39-manylinux2014_x86_64.whl" }
jaxlib = { extras = ["cuda"], version = "0.3.25" }
# jax = "0.3.25"

# Trajectron deps
# trajectron-plus-plus  = {path = "../Trajectron-plus-plus", develop = true}
matplotlib = "^3.5.2"
casadi = "3.5.5"
seaborn = "^0.11.2"
pandas = "^1.4.3"
scipy = "^1.8.1"
pyquaternion = "^0.9.9"
orjson = "^3.7.5"
ncls = "^0.0.65"
dill = "^0.3.5"
pathos = "^0.2.9"
pykalman = "^0.9.5"
scikit-learn = "^1.1.1"
torchmetrics = "0.6.2"

# nuPlan deps
Jinja2 = "^3.1.2"
bokeh = "^2.4.3"
Shapely = "^1.8.4"
geopandas = "^0.11.1"
opencv-python = "<=4.5.1.48"
Pillow = "^9.2.0"
pytorch-lightning = "1.3.8"
selenium = "^4.5.0"
psutil = "^5.9.2"
pyinstrument = "^4.3.0"
boto3 = "1.19.8"
rasterio = "^1.3.2"
SQLAlchemy = " >=1.4.27"
ujson = "^5.5.0"
ray = "^2.0.0"
pyogrio = "^0.4.2"
pyarrow = "^9.0.0"
timm = "^0.6.11"
tensorboard = "2.9.1"
rich = "^12.6.0"

[tool.poetry.dev-dependencies]
black = "22.6.0"
pytest = "^7.1.2"
ptvsd = "^4.3.2"
pre-commit = "^2.20.0"
ipykernel = "^6.16.1"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "main:main_app"
postprocess = "postprocessor:main_app"
expert = "main_expert:main_app"
viz = "viz:main_app"
tools = "tools:cli"

# Task manager
# poetry run poe [options] task_name [task_args]
[tool.poe.tasks]
format = "black ."
isort = "isort --recursive ."
notebook = "poetry run jupyter notebook --no-browser --port=8888"
